Move removeInternalRepeatIdsRecursive() to another file to prevent jest from failing
